// dashboard/src/client/apis/runs-api.ts
export class RunsApi extends BaseAPI {
    /**
     * 
     * @summary Cancel a run
     * @param {string} runId 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public cancelRun(runId: string, options?: RawAxiosRequestConfig) {
        return RunsApiFp(this.configuration).cancelRun(runId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Get run summary
     * @param {string} runId 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public getRun(runId: string, options?: RawAxiosRequestConfig) {
        return RunsApiFp(this.configuration).getRun(runId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Get the immutable workflow snapshot used by this run
     * @param {string} runId 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public getRunDefinition(runId: string, options?: RawAxiosRequestConfig) {
        return RunsApiFp(this.configuration).getRunDefinition(runId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary List runs (paginated)
     * @param {number} [limit] 
     * @param {string} [cursor] 
     * @param {RunStatus} [status] 
     * @param {string} [clientId] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public listRuns(limit?: number, cursor?: string, status?: RunStatus, clientId?: string, options?: RawAxiosRequestConfig) {
        return RunsApiFp(this.configuration).listRuns(limit, cursor, status, clientId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Start a run using the in-memory workflow snapshot
     * @param {RunStartRequest} runStartRequest 
     * @param {string} [idempotencyKey] Optional idempotency key for safe retries; if reused with a different body, return 409
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public startRun(runStartRequest: RunStartRequest, idempotencyKey?: string, options?: RawAxiosRequestConfig) {
        return RunsApiFp(this.configuration).startRun(runStartRequest, idempotencyKey, options).then((request) => request(this.axios, this.basePath));
    }
}