// dashboard/src/client/models/run.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Scheduler Public API (v1)
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ResultError } from './result-error';
// May contain unused imports in some cases
// @ts-ignore
import type { RunArtifact } from './run-artifact';
// May contain unused imports in some cases
// @ts-ignore
import type { RunNodeStatus } from './run-node-status';
// May contain unused imports in some cases
// @ts-ignore
import type { RunStatus } from './run-status';

export interface Run {
    'runId': string;
    'status': RunStatus;
    'definitionHash': string;
    'clientId': string;
    'startedAt'?: string | null;
    'finishedAt'?: string | null;
    'error'?: ResultError;
    'artifacts'?: Array<RunArtifact>;
    'nodes'?: Array<RunNodeStatus>;
}